Reorganized imports and exports, added tailwindParser to exports
