+++ packages/flex/src/lib/index.ts
@@ -1,7 +1,8 @@
-// canvas component
 export { default as Box } from './Box/Box.svelte'
 export { default as Flex } from './Flex/Flex.svelte'
 
-export { createClassParser } from './lib/createClassParser'
-
 export type { NodeProps } from './lib/props'
+
+// parsers
+export { createClassParser } from './parsers/createClassParser'
+export { tailwindParser } from './parsers/tailwindParser'
